+++ src/pages/VolumeLeaderboardPage.js
@@ -5,7 +5,7 @@ import { PageSection } from '../components'
 import PageLayout from '../layouts/PageLayout'
 import SearchWallet from '../components/SearchWallet'
 import VolumeLeaderboard from '../components/Leaderboards/VolumeLeaderboard'
-import VolumeLeaderboardPosition from '../components/VolumeLeaderboardPosition'
+import VolumeLeaderboardPosition from '../components/LeaderBoardPosition/VolumeLeaderboardPosition'
 import { isStarknetAddress } from '../utils'
 import { jediSwapClient } from '../apollo/client'
 import { VOLUME_LEADERBOARD_DATA } from '../apollo/queries'
